+++ src/interfaces/sim800-command-state.enum.ts
@@ -5,3 +5,12 @@ export enum Sim800CommandState {
   Done = 'Done',
   Error = 'Error',
 }
+
+export enum NetworkStatus {
+  NotRegistered = '0',
+  RegisteredHomeNetwork = '1',
+  NotRegisteredSearching = '2',
+  RegistrationDenied = '3',
+  Unknown = '4',
+  RegisteredRoaming = '5',
+}